pyproject.toml
[project]
name = "cubemarsapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aioconsole>=0.8.2",
    "gs-usb>=0.3.0",
    "libusb>=1.0.29.post1",
    "pyserial>=3.5",
    "python-can>=4.6.1",
    "pyusb>=1.3.1",
]

[project.scripts]
cubemars = "tools.cli:main"
